Fix most (according to warnings) remaining 64bit issues in tool code
rvGEWindowWrapper is still TODO - it's not as straightforward, because
it insists on storing a pointer in an idWinVar (using idWinInt), but there
is no idWinVar type for pointers

+++ neo/tools/radiant/MapInfo.cpp
@@ -91,9 +91,9 @@ BOOL CMapInfo::OnInitDialog()
 
   CMapStringToPtr mapEntity;
 
-  int nValue = 0;
-	for (entity_t* pEntity=entities.next ; pEntity != &entities ; pEntity=pEntity->next)
-	{
+  intptr_t nValue = 0;
+  for (entity_t* pEntity=entities.next ; pEntity != &entities ; pEntity=pEntity->next)
+  {
 	m_nTotalEntities++;
 	nValue = 0;
 	mapEntity.Lookup(pEntity->eclass->name, reinterpret_cast<void*&>(nValue));
